# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Northwind Resolve API
  version: 1.0.0
  description: |
    Contract-first API for the Northwind Resolve internal operations platform.
    This document defines the MVP wire contract only. It does not imply that
    persistence, legacy adapters, Gemini, or Solana integrations are available.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/billing/exceptions:
    get:
      tags: [Billing]
      operationId: listBillingExceptions
      summary: List billing quality exceptions
      parameters:
        - $ref: '#/components/parameters/RiskLevelFilter'
        - $ref: '#/components/parameters/BillingExceptionStatusFilter'
        - $ref: '#/components/parameters/RegionFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Billing exception queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingExceptionListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /api/billing/exceptions/{id}:
    get:
      tags: [Billing]
      operationId: getBillingException
      summary: Get a billing exception and review context
      parameters:
        - $ref: '#/components/parameters/BillingExceptionId'
      responses:
        '200':
          description: Billing exception detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingException'
        '404':
          $ref: '#/components/responses/BillingExceptionNotFound'
  /api/billing/exceptions/{id}/review:
    post:
      tags: [Billing]
      operationId: reviewBillingException
      summary: Review a billing exception and optionally record a bill correction
      parameters:
        - $ref: '#/components/parameters/BillingExceptionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingExceptionReviewRequest'
      responses:
        '200':
          description: Review result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingExceptionReviewResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/BillingExceptionNotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
...
components:
  parameters:
...
    BillingExceptionId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
    RegionFilter:
      name: region
      in: query
      schema: { type: string, minLength: 1, maxLength: 64 }
...
    RiskLevelFilter:
      name: riskLevel
      in: query
      schema: { $ref: '#/components/schemas/RiskLevel' }
    BillingExceptionStatusFilter:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/BillingExceptionStatus' }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    PageSize:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  responses:
    InvalidRequest:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
...
    BillingExceptionNotFound:
      description: Billing exception does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    StateConflict:
      description: Request conflicts with the current resource state or a prior idempotent request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, code, message, traceId]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, minimum: 100, maximum: 599 }
        code: { type: string, minLength: 1, maxLength: 100 }
        message: { type: string, minLength: 1, maxLength: 500 }
        traceId: { type: string, minLength: 1, maxLength: 128 }
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
        - type: object
          required: [violations]
          properties:
            violations:
              type: array
              minItems: 1
              items: { $ref: '#/components/schemas/FieldViolation' }
    FieldViolation:
      type: object
      required: [field, message]
      properties:
        field: { type: string, minLength: 1, maxLength: 128 }
        message: { type: string, minLength: 1, maxLength: 500 }
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, minimum: 0, format: int64 }
        totalPages: { type: integer, minimum: 0 }
...
    FieldVisitStatus:
      type: string
      enum: [REQUESTED, SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
    BillingExceptionStatus:
      type: string
      enum: [OPEN, IN_REVIEW, RESOLVED, DISMISSED]
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    BillingReviewAction:
      type: string
      enum: [VERIFY_READING, REQUEST_FIELD_VISIT, CORRECT_BILL, APPROVE]
...
    FieldVisitRequest:
      type: object
      required: [requestedFor, visitReason]
      properties:
        requestedFor: { type: string, format: date-time }
        visitReason: { type: string, minLength: 1, maxLength: 500 }
        meterId: { type: string, maxLength: 64 }
        instructions: { type: string, maxLength: 500 }
      description: Contains only operational details required for field work.
    FieldVisit:
      type: object
      required: [id, caseId, status, scheduledAt]
      properties:
        id: { type: string, format: uuid }
        caseId: { type: string, minLength: 1, maxLength: 64 }
        status: { $ref: '#/components/schemas/FieldVisitStatus' }
        scheduledAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        outcome: { type: string, maxLength: 1000, nullable: true }
...
    BillCorrection:
      type: object
      required: [id, accountId, originalValue, correctedValue, reason, region, createdAt]
      properties:
        id: { type: string, format: uuid }
        accountId: { type: string, minLength: 1, maxLength: 64 }
        originalValue: { type: number, format: double }
        correctedValue: { type: number, format: double }
        reason: { type: string, minLength: 1, maxLength: 500 }
        region: { type: string, minLength: 1, maxLength: 64 }
        createdAt: { type: string, format: date-time }
...
    BillingException:
      type: object
      required: [id, accountId, riskScore, riskLevel, reasonCodes, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        accountId: { type: string, minLength: 1, maxLength: 64 }
        caseId: { type: string, maxLength: 64, nullable: true }
        riskScore: { type: integer, minimum: 0, maximum: 100 }
        riskLevel: { $ref: '#/components/schemas/RiskLevel' }
        reasonCodes:
          type: array
          minItems: 1
          items: { type: string, minLength: 1, maxLength: 100 }
        status: { $ref: '#/components/schemas/BillingExceptionStatus' }
        region: { type: string, minLength: 1, maxLength: 64 }
        createdAt: { type: string, format: date-time }
        reviewedBy: { type: string, maxLength: 128, nullable: true }
    BillingExceptionListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BillingException' }
        page: { $ref: '#/components/schemas/PageMetadata' }
    BillingExceptionReviewRequest:
      oneOf:
        - $ref: '#/components/schemas/VerifyReadingReviewRequest'
        - $ref: '#/components/schemas/RequestFieldVisitReviewRequest'
        - $ref: '#/components/schemas/CorrectBillReviewRequest'
        - $ref: '#/components/schemas/ApproveReviewRequest'
      discriminator:
        propertyName: action
        mapping:
          VERIFY_READING: '#/components/schemas/VerifyReadingReviewRequest'
          REQUEST_FIELD_VISIT: '#/components/schemas/RequestFieldVisitReviewRequest'
          CORRECT_BILL: '#/components/schemas/CorrectBillReviewRequest'
          APPROVE: '#/components/schemas/ApproveReviewRequest'
    ReviewRequestBase:
      type: object
      required: [action, reviewedBy]
      properties:
        action: { $ref: '#/components/schemas/BillingReviewAction' }
        reviewedBy: { type: string, minLength: 1, maxLength: 128 }
        notes: { type: string, maxLength: 1000 }
    VerifyReadingReviewRequest:
      allOf:
        - $ref: '#/components/schemas/ReviewRequestBase'
        - type: object
          properties:
            action: { type: string, enum: [VERIFY_READING] }
    RequestFieldVisitReviewRequest:
      allOf:
        - $ref: '#/components/schemas/ReviewRequestBase'
        - type: object
          required: [fieldVisitRequest]
          properties:
            action: { type: string, enum: [REQUEST_FIELD_VISIT] }
            fieldVisitRequest: { $ref: '#/components/schemas/FieldVisitRequest' }
    CorrectBillReviewRequest:
      allOf:
        - $ref: '#/components/schemas/ReviewRequestBase'
        - type: object
          required: [correction]
          properties:
            action: { type: string, enum: [CORRECT_BILL] }
            correction:
              type: object
              required: [originalValue, correctedValue, reason, region]
              properties:
                originalValue: { type: number, format: double }
                correctedValue: { type: number, format: double }
                reason: { type: string, minLength: 1, maxLength: 500 }
                region: { type: string, minLength: 1, maxLength: 64 }
    ApproveReviewRequest:
      allOf:
        - $ref: '#/components/schemas/ReviewRequestBase'
        - type: object
          properties:
            action: { type: string, enum: [APPROVE] }
    BillingExceptionReviewResponse:
      type: object
      required: [exception, action, reviewedAt]
      properties:
        exception: { $ref: '#/components/schemas/BillingException' }
        action: { $ref: '#/components/schemas/BillingReviewAction' }
        correction: { $ref: '#/components/schemas/BillCorrection', nullable: true }
        fieldVisit: { $ref: '#/components/schemas/FieldVisit', nullable: true }
        reviewedAt: { type: string, format: date-time }
...